+++ src/nationalarchives/analytics.mjs
@@ -402,7 +402,7 @@ class GA4 extends EventTracker {
   }
 }
 
-const ga4Id = document.documentElement.getAttribute("data-ga4");
+const ga4Id = document.documentElement.dataset.ga4;
 if (ga4Id) {
   new GA4({ id: ga4Id });
 }